requirements.txt
asgiref==3.8.0
Django==4.0.6
djangogrpcframework==0.2.1
djangorestframework==3.15.0
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
setuptools==69.2.0
sqlparse==0.4.4
